[DRFT] Sync and Async Producers
